Improved Cygwin .startxwinrc with a sleep forever option
The recent Xinit package update changes the behaviour of startx so that
it will now exit on completion of the .startxwinrc. This means we have
to  run something, but if we just want nothing but an X11 server, we can
just run a sleep command forever. Yes, it's a hack, but it does work.