Build an OpenSSH configuration suitable for the running version
New OpenSSH releases frequently introduce configuration parameters or
options which older versions won't simply ignore. As such, if we want to
make use of these new parameters/options, we need to ensure they're not
present in the configuration when using an older client which won't
understand them.

To handle this, we have a template configuration file for each supported
OpenSSH version containing all of the valid configuration parameters,
which is appended to our custom configuration which we share between all
versions (at least for now). The result is stored as the actual config
which OpenSSH will use. Not ideal, but still fairly elegant & flexible.