Exclude posh/Modules from VS Code search & quick open
This was a PITA to track down. For a while now PowerShellEditorServices
would pin a CPU core on startup when editing a PowerShell script, only
in this repository, eventually settling after ~10 minutes and consuming
~15GB of RAM. While this was happening, auto-formatting of PowerShell
scripts on save would block due to the PowerShell LSP being busy.

It turns out the issue was the PowerShell extension loading every
PowerShell file in the repository for CodeLens support. That includes
everything in the posh/Modules directory (i.e. all user-scope modules).
At the time of writing that's about ~16,700 files, totalling ~6.42GB.
It's also not simply loading them, but keeping the content in memory in
data structures for CodeLens usage. Hence the massive memory usage.

The quick solution is to add the path to the VSC search.exclude setting.