Improve SSH agent socket setup under WSL to work in multiple distros
More specifically, to work in multiple WSL distros which are running
simultaneously. The old approach would only setup the SSH agent socket
in the first distribution which was launched. Subsequent distros would
not have the SSH agent socket set-up while the earlier socket was still
active. The reason is because our detection method was to check if an
existing socket was open in the kernel socket table under the expected
SSH_AUTH_SOCK path.

Under WSL2 distributions share a single Linux kernel instance, but
operate under separate namespaces. However, the network namespace is
shared, and so each running distribution can see the sockets of other
running distros. As such, subsequent launched distributions would see
the first distribution's socket in the kernel socket table, and assume
that only setting and exporting $SSH_AUTH_SOCK was required.

The new approach is to explicitly try and open any existing socket at
the SSH_AUTH_SOCK path, and if it fails, remove the dead socket and
create a new one by launching the appropriate socat instance.

More background information here:
https://github.com/microsoft/WSL/issues/5467