Add Uninstall-ObsoleteModule function
This is verbatim from the script included in PSWinGlue. The issue is
that module is designed for Windows systems, but this function is of use
on any supported PowerShell platform. It's also the only such function
in that module, so I haven't bothered to split it out yet into its own
module, or rework the module itself to be multi-platform.